decouple /username and /username/progress from user.js to make it more modular
